refactor blog iteration

+++ includes/class-fontawesome-deactivator.php
@@ -18,16 +18,9 @@ class FontAwesome_Deactivator {
 	 */
 	public static function deactivate() {
 		if ( is_multisite() && is_network_admin() ) {
-			global $wpdb;
-			$blog_ids = $wpdb->get_col( "SELECT blog_id FROM $wpdb->blogs" );
-			$original_blog_id = get_current_blog_id();
-		
-			foreach ( $blog_ids as $blog_id ) {
-				switch_to_blog( $blog_id );
+			for_each_blog( function( $blog_id ) {
 				self::delete_transients_current_site();
-			}
-
-			switch_to_blog( $original_blog_id );
+			});
 		} else {
 			self::delete_transients_current_site();
 		}
@@ -42,17 +35,10 @@ class FontAwesome_Deactivator {
 	 * Delete options data.
 	 */
 	public static function uninstall() {
-		if ( is_network_admin() ) {
-			global $wpdb;
-			$blog_ids = $wpdb->get_col( "SELECT blog_id FROM $wpdb->blogs" );
-			$original_blog_id = get_current_blog_id();
-		
-			foreach ( $blog_ids as $blog_id ) {
-				switch_to_blog( $blog_id );
+		if ( is_multisite() && is_network_admin() ) {
+			for_each_blog( function( $blog_id ) {
 				self::delete_options_current_site();
-			}
-
-			switch_to_blog( $original_blog_id );
+			});
 		} else {
 			self::delete_options_current_site();
 		}
